functions/python/requirements.txt
firebase-functions>=0.4.2
firebase-admin>=6.2.0
functions-framework>=3.4.0
openai>=1.58.1
python-dotenv>=1.0.0
